+++ graphix/flow/_find_gpflow.py
@@ -583,7 +583,7 @@ def compute_partial_order_layers(correction_matrix: CorrectionMatrix[_AM_co]) ->
     Returns
     -------
     layers : tuple[frozenset[int], ...]
-        Partial order between corrected qubits in a layer form. The frozenset `layers[i]` comprises the nodes in layer `i`. Nodes in layer `i` are "larger" in the partial order than nodes in layer `i+1`. Output nodes are always in layer 0.
+        Partial order between corrected qubits in a layer form. The frozenset `layers[i]` comprises the nodes in layer `i`. Nodes in layer `i` are "larger" in the partial order than nodes in layer `i+1`. Output nodes are always in layer 0 when they exist.
 
     or `None`
         If the correction matrix is not compatible with a partial order on the the open graph, in which case the associated ordering matrix is not a DAG. In the context of the flow-finding algorithm, this means that the input open graph does not have Pauli (or generalised) flow.
@@ -591,8 +591,6 @@ def compute_partial_order_layers(correction_matrix: CorrectionMatrix[_AM_co]) ->
     Notes
     -----
     - The partial order of the Pauli (or generalised) flow :math:`<_c` is the transitive closure of :math:`\lhd_c`, where the latter is related to the ordering matrix :math:`NC` as :math:`v \lhd_c w \Leftrightarrow (NC)_{w,v} = 1`, for :math:`v, w, \in O^c` two non-output nodes of `aog`. The ordering matrix is the product of the order-demand and the correction matrices and it is the adjacency matrix of the directed acyclical graph encoding the partial order.
-
-    - If the open graph has flow, it must have outputs, so `layers[0]` always contains a finite set of nodes.
 
     See Lemma 3.12, and Theorem 3.1 in Mitosek and Backens, 2024 (arXiv:2410.23439).
     """
@@ -602,9 +600,7 @@ def compute_partial_order_layers(correction_matrix: CorrectionMatrix[_AM_co]) ->
     if (topo_gen := _try_ordering_matrix_to_topological_generations(ordering_matrix)) is None:
         return None  # The NC matrix is not a DAG, therefore there's no flow.
 
-    layers = [
-        frozenset(aog.og.output_nodes)
-    ]  # Output nodes are always in layer 0. If the open graph has flow, it must have outputs, so we never end up with an empty set at `layers[0]`.
+    layers = [frozenset(aog.og.output_nodes)] if aog.og.output_nodes else []  # Output nodes are always in layer 0.
 
     # If m >_c n, with >_c the flow partial order for two nodes m, n, then layer(n) > layer(m).
     # Therefore, we iterate the topological sort of the graph in _reverse_ order to obtain the order of measurements.